Ensure that no live auctions exists before remove category.

+++ server/src/models/category/category.service.ts
@@ -1,5 +1,7 @@
 import {
 	BadRequestException,
+	forwardRef,
+	Inject,
 	Injectable,
 	NotFoundException,
 } from '@nestjs/common';
@@ -10,12 +12,14 @@ import { AuctionDocument } from '../auction/schema/auction.schema';
 import { UpdateCategoryDto } from './dto';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { Category, CategoryDocument } from './schema/category.schema';
+import { AuctionsService } from 'src/models/auction/auctions.service';
 
 @Injectable()
 export class CategoryService {
 	constructor(
 		@InjectModel(Category.name)
 		private readonly categoryModel: Model<CategoryDocument>,
+		private readonly auctionService: AuctionsService,
 	) {}
 
 	/**
@@ -126,6 +130,16 @@ export class CategoryService {
 	 * @param categoryId: string
 	 */
 	async remove(categoryId: string) {
+		//? Ensure before remove that there is no upcoming or ongoing auctions related to this category
+		const isExists =
+			this.auctionService.isThereAnyRunningAuctionRelatedToCategory(categoryId);
+		if (isExists) {
+			throw new BadRequestException(
+				'There are ongoing or upcoming auctions related to this category ❌.',
+			);
+		}
+
+		//* Find the category and delete it
 		const category = await this.categoryModel.findByIdAndRemove(categoryId);
 
 		if (!category) throw new NotFoundException('Category not found ❌');
